+++ bmds/logic/defaults.py
@@ -1,6 +1,0 @@
-from . import rules
-from .. import constants
-
-
-def get_default_rules(dtype):
-