+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 v0.4.0 (NEXT)
 -------------------
 
+* Added Dichotomous-Hill model to list of dichotomous models
 * Allowed global overrides (such as BMR settings) as an optional input to ``Session.add_default_models``
 * Updated test-logic outputs for individual tests (and added tests)