+++ HISTORY.rst
@@ -9,6 +9,9 @@ v0.4.0 (NEXT)
 * Allowed global overrides (such as BMR settings) as an optional input to ``Session.add_default_models``
 * Updated test-logic outputs for individual tests (and added tests)
 * For continuous summary datasets, rename ``responses`` array to ``means``
+* By default, polynomial-like models are run multiple times with different degrees.
+    - Previously, a single polynomial model was added with an order of ``min(n-1, 8)``, where ``n`` is the number of dose-groups. Now, multilpe models are added ranging from ``[3 - min(n-1, 8)]``, inclusive
+    - Polynomial like models include: Polynomial, Multistage, and Multistage-Cancer
 
 
 v0.3.0 (2016-12-05)